Adapt program for t_minishell structure and handle list conversion, memory leaks, and bug fix
Changes:

The program has been adapted to work with the t_minishell structure.
The parse_arguments function now returns a list of type t_input.
Implemented a function to convert the list into a 2D array.
Memory leaks in the new parser are now handled, except in cases involving pipes or redirection.
All changes made adhere to norminnete guidelines.
Fixed a bug in the echo function that caused it to not count the last argument correctly.